requirements.txt
azure-mgmt-resource 
azure-mgmt-sql 
azure-identity